refactor: allow custom condition and action
This commit refactors the AutoResponseTrigger class to add functionality for custom condition and action functions to the trigger response. The setCustomCondition and setCustomAction methods were added to allow for setting these functions externally. Also, two new methods were added for the custom functions, customCondition and customAction. The commit also removes unused code and improves error handling by throwing an error if a source or bot message is not set.